saasa

+++ application/jadwal_guru_piket.php
@@ -90,7 +90,6 @@
                 <th>Hari</th>
                 <th>Guru</th>
                 <th>Tanggal</th>
-
                 <th>Action</th>
               </tr>
             </thead>
@@ -102,18 +101,22 @@
 
 
               $no = 1;
-              while ($r = mysql_fetch_array($tampil)) {
-                echo "<tr><td>$no</td>
-                              <td>$r[nip]</td>
-                              <td>$r[hari]</td>
-                              <td>$r[nama_guru]</td>
-                              <td>" . tgl_indo($r['tanggal']) . "</td>";
-                echo "<td style='width:80px !important'><center>
-                                        <a class='btn btn-success btn-xs' title='Lihat Journal' href='index.php?view=jadwalgurupiket&act=edit&nip=$r[nip]'><span class='glyphicon glyphicon-pencil'></span> Edit</a>
-                                        <a class='btn btn-danger btn-xs' title='Hapus Jadwal' href='index.php?view=jadwalgurupiket&act=delete&nip=$r[nip]' onclick=\"return confirm('Apakah Anda yakin ingin menghapus jadwal ini?')\"><span class='glyphicon glyphicon-trash'></span> Delete</a>
-                                      </center></td>";
-                echo "</tr>";
-                $no++;
+              if (mysql_num_rows($tampil) > 0) { // Memeriksa apakah ada data
+                while ($r = mysql_fetch_array($tampil)) {
+                  echo "<tr><td>$no</td>
+                                <td>$r[nip]</td>
+                                <td>$r[hari]</td>
+                                <td>$r[nama_guru]</td>
+                                <td>" . tgl_indo($r['tanggal']) . "</td>";
+                  echo "<td style='width:80px !important'><center>
+                                          <a class='btn btn-success btn-xs' title='Lihat Journal' href='index.php?view=jadwalgurupiket&act=edit&nip=$r[nip]'><span class='glyphicon glyphicon-pencil'></span> Edit</a>
+                                          <a class='btn btn-danger btn-xs' title='Hapus Jadwal' href='index.php?view=jadwalgurupiket&act=delete&nip=$r[nip]' onclick=\"return confirm('Apakah Anda yakin ingin menghapus jadwal ini?')\"><span class='glyphicon glyphicon-trash'></span> Delete</a>
+                                        </center></td>";
+                  echo "</tr>";
+                  $no++;
+                }
+              } else {
+                echo "<tr><td colspan='6' style='text-align:center;'>Tidak ada data</td></tr>"; // Menampilkan pesan jika tidak ada data
               }
               ?>
             <tbody>
